+++ src/sora/sora_client.cpp
@@ -146,8 +146,15 @@ void SoraClient::Connect() {
 
     std::shared_ptr<Websocket> ws;
     if (ssl) {
-      ws.reset(new Websocket(Websocket::ssl_tag(), ioc_, config_.insecure,
-                             config_.client_cert, config_.client_key));
+      if (config_.proxy_url.empty()) {
+        ws.reset(new Websocket(Websocket::ssl_tag(), ioc_, config_.insecure,
+                               config_.client_cert, config_.client_key));
+      } else {
+        ws.reset(new Websocket(Websocket::https_proxy_tag(), ioc_,
+                               config_.insecure, config_.client_cert,
+                               config_.client_key, config_.proxy_url,
+                               config_.proxy_username, config_.proxy_password));
+      }
     } else {
       ws.reset(new Websocket(ioc_));
     }
@@ -276,9 +283,7 @@ void SoraClient::DoSendConnect(bool redirect) {
     json_message["redirect"] = true;
   }
 
-  if (config_.multistream) {
-    json_message["multistream"] = true;
-  }
+  json_message["multistream"] = config_.multistream;
 
   if (config_.simulcast) {
     json_message["simulcast"] = true;